+++ packages/web/src/app/doc-pages/components/forms/segmented-controls-doc/segmented-controls-doc.component.ts
@@ -17,7 +17,7 @@ export class SegmentedControlsDocComponent {
   ];
 
   exampleOverview =
-    `<div class="e-segmented-controls">
+    `<div class="e-segmented-controls" style="width: 300px">
   <input type="radio" name="so-1-1" id="so-1-1" checked>
   <input type="radio" name="so-1-1" id="so-1-2">
   <label for="so-1-1" data-value="Option 1"><span class="e-segmented-controls__title">` +
@@ -30,7 +30,7 @@ export class SegmentedControlsDocComponent {
 `;
 
   example1 =
-    `<div class="e-segmented-controls">
+    `<div class="e-segmented-controls" style="width: 300px">
   <input type="radio" name="sc-1-1" id="sc-1-1" checked>
   <input type="radio" name="sc-1-1" id="sc-1-2">
   <label for="sc-1-1" data-value="Option 1"><span class="e-segmented-controls__title">` +
@@ -43,58 +43,76 @@ export class SegmentedControlsDocComponent {
 `;
 
   example2 =
-    `<div class="e-segmented-controls">
+    `<div class="e-segmented-controls" style="width: 350px">
   <input type="radio" name="sc-1-2" id="sc-1-1-1" checked>
   <input type="radio" name="sc-1-2" id="sc-1-1-2">
   <input type="radio" name="sc-1-2" id="sc-1-1-3">
-  <label for="sc-1-1-1" data-value="Option 1"><span class="e-segmented-controls__title">` +
+    <label for="sc-1-1-1" data-value="Option 1">
+      <span class="e-segmented-controls__title">` +
     exampleContents.words.views['eng-GBR'][0] +
-    `</span></label>
-  <label for="sc-1-1-2" data-value="Option 2"><span class="e-segmented-controls__title">` +
+    `</span>
+    </label>
+    <label for="sc-1-1-2" data-value="Option 2">
+      <span class="e-segmented-controls__title">` +
     exampleContents.words.views['eng-GBR'][1] +
-    `</span></label>
-  <label for="sc-1-1-3" data-value="Option 3"><span class="e-segmented-controls__title">` +
+    `</span>
+    </label>
+    <label for="sc-1-1-3" data-value="Option 3">
+      <span class="e-segmented-controls__title">` +
     exampleContents.words.views['eng-GBR'][2] +
-    `</span></label>
+    `</span>
+    </label>
 </div>
 `;
 
   exampleLarge =
-    `<div class="e-segmented-controls e-segmented-controls--lg">
+    `<div class="e-segmented-controls e-segmented-controls--lg" style="width: 300px">
   <input type="radio" name="sclg-1-1" id="sclg-1-1" checked>
   <input type="radio" name="sclg-1-1" id="sclg-1-2">
-  <label for="sclg-1-1" data-value="Option 1"><span class="e-segmented-controls__title">` +
+  <label for="sclg-1-1" data-value="Option 1">
+    <span class="e-segmented-controls__title">` +
     exampleContents.words.views['eng-GBR'][0] +
-    `</span></label>
-  <label for="sclg-1-2" data-value="Option 2"><span class="e-segmented-controls__title">` +
+    `</span>
+  </label>
+  <label for="sclg-1-2" data-value="Option 2">
+    <span class="e-segmented-controls__title">` +
     exampleContents.words.views['eng-GBR'][1] +
-    `</span></label>
+    `</span>
+  </label>
 </div>
 `;
 
   exampleMedium =
-    `<div class="e-segmented-controls">
+    `<div class="e-segmented-controls" style="width: 300px">
   <input type="radio" name="scmd-1-1" id="scmd-1-1" checked>
   <input type="radio" name="scmd-1-1" id="scmd-1-2">
-  <label for="scmd-1-1" data-value="Option 1"><span class="e-segmented-controls__title">` +
+  <label for="scmd-1-1" data-value="Option 1">
+    <span class="e-segmented-controls__title">` +
     exampleContents.words.views['eng-GBR'][0] +
-    `</span></label>
-  <label for="scmd-1-2" data-value="Option 2"><span class="e-segmented-controls__title">` +
+    `</span>
+  </label>
+  <label for="scmd-1-2" data-value="Option 2">
+    <span class="e-segmented-controls__title">` +
     exampleContents.words.views['eng-GBR'][1] +
-    `</span></label>
+    `</span>
+  </label>
 </div>
 `;
 
   exampleSmall =
-    `<div class="e-segmented-controls e-segmented-controls--sm" style="max-width: 280px">
-  <input type="radio" name="scsm-1-1" id="scsm-1-1" checked>
-  <input type="radio" name="scsm-1-1" id="scsm-1-2">
-  <label for="scsm-1-1" data-value="Option 1"><span class="e-segmented-controls__title">` +
+    `<div class="e-segmented-controls e-segmented-controls--sm" style="width: 300px">
+  <input type="radio" name="scsm-1-2" id="scsm-1-1-1" checked>
+  <input type="radio" name="scsm-1-2" id="scsm-1-1-2">
+    <label for="scsm-1-1-1" data-value="Option 1">
+      <span class="e-segmented-controls__title">` +
     exampleContents.words.views['eng-GBR'][0] +
-    `</span></label>
-  <label for="scsm-1-2" data-value="Option 2"><span class="e-segmented-controls__title">` +
+    `</span>
+    </label>
+    <label for="scsm-1-1-2" data-value="Option 2">
+      <span class="e-segmented-controls__title">` +
     exampleContents.words.views['eng-GBR'][1] +
-    `</span></label>
+    `</span>
+    </label>
 </div>
 `;
 }